requirements.txt
mcp==1.24.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
httpx
python-dotenv>=1.0.0

